items.py created, registering users can be stored on sql database
