+++ code/security.py
@@ -14,10 +14,13 @@ usernameMapping = {u.username : u for u in users}
 userIdMapping = {u.id : u for u in users}
 
 def authenticate(username:str, password:str):
-    user = usernameMapping.get(username, None)
+    user = User.findByUsername(username)
     if user and user.password == password:
         return user
 
 def identity(payload):
     userId = payload['identity']
-    return userIdMapping.get(userId, None)
+    return User.findById(userId)
+
+
+
