+++ code/user.py
@@ -1,7 +1,6 @@
 import sqlite3
 
-from colorama import Cursor
-
+from flask_restful import Resource, reqparse
 
 
 class User:
@@ -48,3 +47,34 @@ class User:
         connection.close()
 
         return user
+
+class UserRegister(Resource):
+
+    parser = reqparse.RequestParser()
+    parser.add_argument('username',
+        type = str,
+        required = True,
+        help = 'This field cannot be left blank'
+    )
+    parser.add_argument('password',
+        type = str,
+        required = True,
+        help = 'This field cannot be left blank'
+    )
+    def post(self):
+        data = UserRegister.parser.parse_args()
+
+        if User.findByUsername(data['username']):
+            return {'message': 'User with that username already exists'}, 400
+
+        connection = sqlite3.connect('data.db')
+        cursor = connection.cursor()
+        
+        query = 'INSERT INTO users VALUES (Null,?,?)'
+        cursor.execute(query, (data['username'], data['password'])) 
+
+        connection.commit()
+
+        connection.close()
+
+        return {'message': 'User successfully created'}, 201
